requirements.txt
SQLAlchemy~=2.0.41
pytz~=2025.2
aiogram~=3.21.0
APScheduler~=3.11.0
python-dotenv~=1.1.1
greenlet
aiosqlite